Add files via upload
This project combines two different data files regarding user signups to use a particular product. A Light Gradient Boosting Machine Classifier is used to determine which features are more important to the target which is becoming an "adopted" user.  LightGBM was used here because there may not have been enough observations for XGBoost.